chore: fix false Vale linter errors due to incorrect config
The `BlockIgnores` and `TokenIgnores` regular expressions in the Vale config were causing incorrect / unnecesary modifications to input AsciiDoc files, and that lead to false linter errors.

There is no need to try to exclude special asciidoc syntax (like `image:` and `include:` macros and `link:` and `<<>>` links) for Vale.

Before linting the text, Vale passes the asciidoc files through an asciidoc processor and works on the HTML output. This means that all special asciidoc syntax that does not get rendered in HTML is ignored by Vale (by design).

Also, when working with asciidoc files Vale automatically processes the front matter, so there is no need to explicitly mention it in `BlockIgnores`.