docs: improve and simplify showing errors in Upload demos
Use `showErrorNotification` helper from org.vaadin.artur:a-vaadin-helper
to simplify showing notifications. This also makes it so that multiple
errors can be shown at the same time when `fileRejectHandler` is called
multiple times for a single action. For example when the user selects
more than 3 files when `maxFiles` is 3, there will be an error shown for
each file that goes over the limit. Before this change this would only
show the latest error since each call was overriding the previous error
message.