Change DSP installation instructions to use npx
Installing vaadin-cli seems to be a hassle for many users not used to npm, e.g. with them not having the npm/bin directory in their PATH, which means the vaadin command doesn't work, etc. We can avoid all that hassle by instructing them to use the npx command instead, which frees you of the installation step entirely. Thus we don't even have to mention vaadin-cli as a separate prerequisite.