Rewrote the intro to routing and navigation
New version, even though being longer, should better explain why we have the concept in the first place, where it should be used and when it should not be used.

Also refactored the docs a bit. Probably getting close to tackling #1482 as well with this pull request (don't have good examples for people not using 23.2-SNAPSHOT yet, but who wouldn't 😎).