Update index.asciidoc
I made some basic edits, to include some additional lead-in text. However, it seems fairly spare. I think it needs more text at the start of each minor section.

Also, the end of one of the sentences seems odd to me:  "which allows the codebase to be transitioned gradually into Vaadin upgrading single views."  It may be that I don't know the product, but that wording seems odd.  Maybe it should be, "which allows the codebase to be transitioned gradually into Vaadin "by" upgrading single views"?  I feel like something is missing.  It may just be my lack of undersanding and not the text, so I left it alone.