Update flow-configuration.asciidoc
I made a few edits and added some text.  Again, this seems spare to me, but it may be that I don't understand where it will appear.  If it's meant to be text that will appear when using the software, brevity is fine. But if it's in the main documentation on the web, it might seem more substantial with more text and a bit more clarification.