Docs: add upgrade notes for Hilla 2.0
Fixes vaadin/hilla#784
Fixes vaadin/hilla#777
Fixes vaadin/hilla#778

This adds the upgrade guide section for Hilla 2.0.

Much of the existing content in the guide applies to Hilla applications
as well, and the changes elsewhere in Vaadin 24.0 (in the base stack and Vaadin
components) have significant impact on migrating Hilla apps to 2.0.

It seems better to add a Hilla section here rather then an entirely separate
guide in Hilla docs, to avoid duplicated content and making it more
convenient for Hilla users to follow.