Update e2e-testing-with-testbench.adoc
In this update, a minor typo error in the message "Vaadin TestBench contains handy base classes that you can use aa a bases for your e2e tests" has been corrected. The word "aa" has been replaced with "as", resulting in the correct message "Vaadin TestBench contains handy base classes that you can use as a bases for your e2e tests". This correction improves the clarity and accuracy of the message, making it easier for readers to understand and use the information provided.